% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-display.R
\name{build_display}
\alias{build_display}
\title{Build a trelliscope display of flyover plots.}
\usage{
build_display(
  plot_data,
  display_name,
  output_dir,
  nrow = 1,
  ncol = 2,
  sort_var = "variable",
  sort_direction = "asc",
  self_contained = FALSE,
  ...
)
}
\arguments{
\item{plot_data}{A \code{tibble} output from \code{build_plots}.}

\item{display_name}{Character string of the name to give the trelliscope display.}

\item{output_dir}{The file path of the directory to create in which the
outputs of the display will be written.  Caution: the function
will overwrite any files stored in this directory!}

\item{nrow}{The number of rows of plots to display.}

\item{ncol}{The number of columns of plots to display.}

\item{sort_var}{Character string of the cognostic to sort on.  Defaults to the
variable name, but can be any of the cognostics of interest.}

\item{sort_direction}{Character string specifying direction of sort;
can be "asc" for ascending or "desc" for descending.
Defaults to "asc".}

\item{self_contained}{Logical; whether to render the display as a
self-contained HTML document.  This should be set
to \code{TRUE} for including the display in an
R markdown HTML document, but generally can be left
as \code{FALSE} for producing an output directory
where plots and relevant files will be stored.}

\item{...}{additional arguments passed to \code{trelliscopejs::trelliscope}.}
}
\description{
This function produces an interactive display for collecting,
organizing, and displaying the plots generated in \code{\link{build_plots}}.
The bulk of the work is done by \code{trelliscopejs::trelliscope}.
Please see the documentation of this package for additional details.
}
