% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-plots.R
\name{build_plots}
\alias{build_plots}
\title{Create plot objects for the trelliscope display.}
\usage{
build_plots(
  stack,
  plot_fun,
  group_var = "flyover_id_",
  keep_type = NULL,
  plot_mods = NULL,
  ...
)
}
\arguments{
\item{stack}{Tabular data that inherits from \code{data.frame}.
This is recommended to be the output of
\code{\link{stack_data}} but can be built by the user.
It is coerced to a \code{tibble}.}

\item{plot_fun}{The \code{flyover} plotting function to apply to the
data. The user may also supply a custom function,
but must be careful to match that function with
the appropriate \code{keep_type} argument.
Note that this function must be passed without parentheses.}

\item{group_var}{Character string; the column name that represents the
source of the data.  It will be used as a grouping
variable in the subsequent plots.}

\item{keep_type}{Depending on the type of plot desired,
only numeric or categorical data can be used.
By default the column type is determined by the
\code{flyover} plotting function passed to the
\code{plot_fun} argument.  If a \code{flyover} plot
is passed, this argument is ignored. However, if the user
specifies a custom plotting function, the
\code{keep_type} argument must be set to one of
\code{"numeric"} or \code{"categorical"}.  Passing
\code{"categorical"} will keep character, factor,
and logical column types.}

\item{plot_mods}{list containing additional layers to the \code{ggplot2} call
such as theme changes, different color scales, etc.
Each layer should be a separate list element.
See \code{?ggplot2::`+.gg`} for more details.}

\item{...}{Additional arguments to pass to the \code{geom} of the
supplied \code{flyover} plot function.  Use this for further
modifications to the plots if needed.}
}
\value{
A \code{tibble} containing, for each relevant variable of the
input data, a row with a plot object and data frame of
cognostics for the trelliscope display. The tibble can
then be passed to \code{build_display} to create the
trelliscope output.
}
\description{
This function allows the user to choose from a menu
of prepackaged plotting styles (or to pass your own) that are designed
to facilitate data comparison between groups.  Plot functions are
specific to one type of data (numeric or categorical) and will
only use those columns when building plots.  For example,
histograms are only useful for numeric data, so categorical
type data is ignored.
}
