% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist-data.R
\name{enlist_data}
\alias{enlist_data}
\title{Create a named list from data}
\usage{
enlist_data(..., names)
}
\arguments{
\item{...}{Comma-separated input of tabular-type data.  Any class
that inherits from \code{data.frame} is permitted.}

\item{names}{Character vector of names corresponding to each data
data element in \code{...}}
}
\value{
A named list of \code{tibble}s
}
\description{
This is a convenience function serving as an
intermediate step in preparing the data for visualization.
The function builds a named list of the data elements.
The names will be used to group the data in subsequent
visualizations, so please ensure they are meaningful.
Each element will be coerced to a \code{tibble}
and row names are dropped, SORRY.
}
\examples{
x <- data.frame(a = 1:3, b = letters[1:3])
y <- data.frame(a = 4:6, b = letters[4:6])
enlist_data(x, y, names = c("old", "new"))

}
