% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyover.R
\docType{package}
\name{flyover-package}
\alias{flyover}
\alias{flyover-package}
\title{flyover: Visually Study High-Level Differences in Data Sets}
\description{
Simplify visual data set comparison!
}
\author{
\strong{Maintainer}: Adam Austin \email{ataustin@gmail.com}

}
