% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyover-histogram.R
\name{flyover_histogram}
\alias{flyover_histogram}
\title{Create overlaid histograms of a numeric column grouped by a specified variable.}
\usage{
flyover_histogram(tbl, var, group_var)
}
\arguments{
\item{tbl}{A \code{tibble} containing the variable of interest and the grouping variable.}

\item{var}{Character string; the numeric variable whose distribution is of interest}

\item{group_var}{Character string; the variable defining different groups.}
}
\value{
A plot object generated by the \code{ggplot2} package.
}
\description{
This function is designed to be called by \code{\link{build_plots}}
but is exported for convenience so the user can test the output.
The function builds a somewhat opinionated histogram in which
distributions can be visually compared across groups.
It is recommended that no more than three groups be compared
with this function to avoid visual clutter.
}
\examples{
\dontrun{
x <- tibble::tibble(group = rep(c("old", "new"), each = 100),
                    rand  = c(rnorm(100), rnorm(100, mean = 1)))
flyover_histogram(x, "rand", "group")
}

}
