% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist-data.R
\name{enlist_data}
\alias{enlist_data}
\title{Create a named list from data}
\usage{
enlist_data(..., names)
}
\arguments{
\item{...}{Data frames, tibbles, or data.tables separated by commas}

\item{names}{Character vector of names corresponding to each data
data element in \code{...}}
}
\value{
A named list
}
\description{
This is a convenience function serving as an
intermediate step in preparing the data for visualization.
The function builds a named list
of data frames (or similar objects).  The names will be used to
refer to the data in subsequent visualizations, so ensure they
are meaningful.
}
\examples{
x <- data.frame(a = 1:3, b = letters[1:3])
y <- data.frame(a = 4:6, b = letters[4:6])
enlist_data(x, y)

}
