% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyover-plots.R
\name{flyover_histogram}
\alias{flyover_histogram}
\alias{flyover_density}
\alias{flyover_binline}
\alias{flyover_bar_dodge}
\alias{flyover_bar_fill}
\title{Minimal functions for basic plotting of groupwise comparisons.}
\usage{
flyover_histogram(tbl, var, group_var)

flyover_density(tbl, var, group_var)

flyover_binline(tbl, var, group_var)

flyover_bar_dodge(tbl, var, group_var)

flyover_bar_fill(tbl, var, group_var)
}
\arguments{
\item{tbl}{A \code{tibble} containing the variable of interest and the grouping variable.}

\item{var}{Character string; the variable name whose distribution is of interest.}

\item{group_var}{Character string; the variable name defining different groups.}
}
\value{
A plot object generated by the \code{ggplot2} package.
}
\description{
These functions are designed to be called by \code{\link{build_plots}}
as part of the flyover workflow, but can also be used as standalone plotting
utilities.  Each function has identical arguments, but each will produce
a different visual comparison of data across groups depeding on the
data type to be compared and the geom selected.  The output
can by modified by directly adding \code{ggplot2} layers in the usual way.
}
\examples{
\dontrun{
x <- tibble::tibble(group = rep(c("old", "new"), each = 100),
                    rand  = c(rnorm(100), rnorm(100, mean = 1)))
flyover_histogram(x, "rand", "group")

y <- tibble::tibble(group = rep(c("old", "new"), each = 100),
                    rand  = sample(letters[1:4], 200, replace = TRUE))
flyover_bar_fill(y, "rand", "group")
}

}
