% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack-data.R
\name{stack_data}
\alias{stack_data}
\title{Stacks a \emph{named} list of tabular data into a single data structure.}
\usage{
stack_data(data_list, drop_mismatches = FALSE, group_var = "flyover_id_")
}
\arguments{
\item{data_list}{A \emph{named} list of tabular data elements that
inherit from \code{data.frame}.
This is generally the output of \code{\link{enlist_data}}.}

\item{drop_mismatches}{Logical.  Whether to drop columns across data
sets if they fail to appear in any one data
set in the list.}

\item{group_var}{Character string.  After stacking, the resulting
data will have a new column with this name.  The values
of this column are the names of \code{data_list}.
These are used to identify the source of data and to
group variables for plotting in downstream functions.}
}
\value{
A single \code{tibble} with a new column having the name
passed to \code{group_var}.
}
\description{
This is a convenience function serving as an
intermediate step in preparing the data for visualization.
It is a basic wrapper for \code{dplyr::bind_rows} that works
with the data preparation flow of \code{flyover}.  The input list
must be named to correctly identify the data sets after stacking.
}
\examples{
x <- list(old = data.frame(a = 1:3),
          new = data.frame(a = 4:6, b = 7:9))
stack_data(x)
stack_data(x, drop_mismatches = TRUE)

}
