% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack-data.R
\name{stack_data}
\alias{stack_data}
\title{Stacks a \emph{named} list of tabular data into a single data structure.}
\usage{
stack_data(data_list, drop_mismatches = FALSE)
}
\arguments{
\item{data_list}{A \emph{named} list of tabular data elements.
This is generally the ouptut of \code{enlist_data}.}

\item{drop_mismatches}{Logical.  Whether to drop columns across data
sets if they fail to appear in any one data
set in the list.}
}
\value{
A single data set; the class of the result obeys the rules of
\code{\link{dplyr::bind_rows}}.  This object will contain a
new column \code{flyover_id} for identifying individual source
data sets after stacking.
}
\description{
This is a convenience function serving as an
intermediate step in preparing the data for visualization.
It is a basic wrapper for \code{dplyr::bind_rows} that works
with the data preparation flow of \code{flyover}.  The input list
must be named to correctly identify the data sets after stacking.
}
\examples{
x <- list(old = data.frame(a = 1:3),
          new = data.frame(a = 4:6, b = 7:9))
stack_data(x)
stack_data(x, drop_mismatches = TRUE)
}
